Add solution for problem `G - Short Task` Link: https://codeforces.com/contest/1512/problem/G Tags: brute force, dp, math, number theory Ref: https://codeforces.com/contest/1512/submission/112580875

+++ codeforces/1512/G.cpp
@@ -2,7 +2,7 @@
 #pragma comment (linker,"/STACK:16777216")
 #include "bits/stdc++.h"
 using namespace std;
-typedef long long ll;
+typedef int ll;
 typedef std::vector<int> vi;
 typedef std::vector<ll> vll;
 #define endl "\n"
